Terraform deployment support for GCP FTDv cluster feature
