--- cluster/gcp/cluster-function/requirements.txt
@@ -1,10 +1,0 @@
-# Function dependencies, for example:
-# package>=version
-google-api-python-client==2.0.2
-google-auth==1.28.0
-google-auth-httplib2==0.1.0
-paramiko==2.11.0
-oauth2client==3.0.0
-google-cloud-secret-manager==2.3.0
-requests==2.25.1
-urllib3==1.26.9
