+++ cluster/gcp/cluster_deployment/modules/cluster_function/scalein-action/requirements.txt
@@ -1,0 +1,10 @@
+# Function dependencies, for example:
+# package>=version
+google-api-python-client==2.170.0
+google-auth==2.40.2
+google-auth-httplib2==0.2.0
+oauth2client==4.1.3
+google-cloud-secret-manager==2.23.3
+paramiko==3.4.0
+requests==2.32.3
+urllib3==2.4.0
