--- cluster/gcp/cluster_function_infra.jinja.schema
@@ -1,10 +1,0 @@
-info:
-  title: cluster_function_infra schema file
-  description: Describes the specifications of the fields in cluster_function_infra.jinja
-  version: 1.0
-imports:
-  - path: cluster_function_infra.jinja 
-properties:
-  resourceNamePrefix:
-    pattern: ^[a-zA-Z0-9]+$
-    description: resourceNamePrefix restricted to single word
