+++ cluster/gcp/infra_deployment/infrastructure_params.tfvars
@@ -1,0 +1,12 @@
+project_id = "<project-id>"                     # GCP project ID (e.g. test-project-12345)
+resource_name_prefix = "<resource-name>"        # Resource prefix (lowercase, e.g. ftdv)
+region = "<region>"                             # GCP region (e.g. us-central1)
+mgmt_ip_cidr_range = ""                         # Management subnet (e.g. 10.112.0.0/27)
+vpc_connector_ip_cidr_range = ""                # VPC connector /28 (e.g. 10.112.50.0/28)
+with_diagnostic = false                         # Enable diagnostic subnet (true/false)
+diag_ip_cidr_range = ""                         # Diagnostic subnet (e.g. 10.112.19.0/27)
+inside_ip_cidr_range = ""                       # Inside subnet (e.g. 10.112.1.0/27)
+outside_ip_cidr_range = ""                      # Outside subnet (e.g. 10.112.2.0/27)
+ccl_ip_cidr_range = ""                          # CCL subnet, /27 recommended (e.g. 10.112.100.0/27)
+ftd_reg_via_public_ip = true                    # Register FTDv via public IP (true/false)
+
