--- cluster/gcp/infrastructure.jinja.schema
@@ -1,10 +1,0 @@
-info:
-  title: infrastructure schema file
-  description: Describes the specifications of the fields in infrastructure.jinja
-  version: 1.0
-imports:
-- path: infrastructure.jinja  
-properties:
-  resourceNamePrefix:
-    pattern: ^[a-zA-Z0-9]+$
-    description: resourceNamePrefix restricted to single word
